requirements.txt
ursina==5.2.0
torch==2.1.0
numpy==1.24.3
gymnasium==0.29.1
transformers==4.35.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
requests==2.31.0
beautifulsoup4==4.12.2
wikipedia==1.4.0
googlesearch-python==1.2.3
openai==1.3.5
pyinstaller==6.2.0
